MWPW-155609: Moved creation of caas url to utils.js
