+++ edsdme/scripts/utils.js
@@ -200,6 +200,82 @@ function preloadLit(miloLibs) {
   document.head.appendChild(preloadLink);
 }
 
+export function getCaasUrl(block) {
+  const domain = 'https://www.adobe.com/chimera-api';
+  const api = new URL(`${domain}/collection?originSelection=dme-partners&draft=false&debug=true&flatFile=false&expanded=true`);
+  return setApiParams(api, block);
+}
+
+function extractTableCollectionTags(el) {
+  let tableCollectionTags = [];
+  Array.from(el.children).forEach((row) => {
+    const cols = Array.from(row.children);
+    const rowTitle = cols[0].innerText.trim().toLowerCase().replace(/ /g, '-');
+    const colsContent = cols.slice(1);
+    if (rowTitle === 'collection-tags') {
+      const [collectionTagsEl] = colsContent;
+      const collectionTags = Array.from(collectionTagsEl.querySelectorAll('li'), (li) => `"${li.innerText.trim().toLowerCase()}"`);
+      tableCollectionTags = [...tableCollectionTags, ...collectionTags];
+    }
+  });
+
+  return tableCollectionTags;
+}
+
+function setApiParams(api, block) {
+  const { el, collectionTag, ietf } = block;
+  const complexQueryParams = getComplexQueryParams(el, collectionTag);
+  if (complexQueryParams) api.searchParams.set('complexQuery', complexQueryParams);
+
+  const [language, country] = ietf.split('-');
+  if (language && country) {
+    api.searchParams.set('language', language);
+    api.searchParams.set('country', country);
+  }
+  return api.toString();
+}
+
+function getComplexQueryParams(el, collectionTag) {
+  const portal = getCurrentProgramType();
+  if (!portal) return;
+
+  const portalCollectionTag = `"caas:adobe-partners/${portal}"`;
+  const tableTags = extractTableCollectionTags(el);
+  const collectionTags = [collectionTag, portalCollectionTag, ...tableTags];
+
+  const partnerLevelParams = getPartnerLevelParams(portal);
+  const partnerRegionParams = getPartnerRegionParams(portal);
+
+  const collectionTagsStr = collectionTags.filter((e) => e.length).join('+AND+');
+  let resulStr = `(${collectionTagsStr})`;
+  if (partnerRegionParams) resulStr += `+AND+${partnerRegionParams}`;
+  if (partnerLevelParams) resulStr += `+AND+${partnerLevelParams}`;
+  // eslint-disable-next-line consistent-return
+  return resulStr;
+}
+
+function getPartnerLevelParams(portal) {
+  const partnerLevel = getPartnerDataCookieValue(portal, 'level');
+  const partnerTagBase = `"caas:adobe-partners/${portal}/partner-level/`;
+  return partnerLevel ? `(${partnerTagBase}${partnerLevel}"+OR+${partnerTagBase}public")` : `(${partnerTagBase}public")`;
+}
+
+function getPartnerRegionParams(portal) {
+  const permissionRegion = getPartnerDataCookieValue(portal, 'permissionregion');
+  const regionTagBase = `"caas:adobe-partners/${portal}/region/`;
+
+  if (!permissionRegion) return `(${regionTagBase}worldwide")`;
+
+  const regionTags = [];
+
+  permissionRegion.split(',').forEach((region) => {
+    const regionValue = region.trim().replaceAll(' ', '-');
+    if (regionValue) regionTags.push(`${regionTagBase}${regionValue}"`);
+  });
+
+  return regionTags.length ? `(${regionTags.join('+OR+')})` : `(${regionTagBase}worldwide")`;
+}
+
 export async function preloadResources(locales, miloLibs) {
   const locale = getLocale(locales);
   const cardBlocks = { 'announcements': '"caas:adobe-partners/collections/announcements"' };
@@ -216,8 +292,7 @@ export async function preloadResources(locales, miloLibs) {
       collectionTag: value,
       ietf: locale.ietf,
     };
-    const { default: PartnerCards } = await import('../components/PartnerCards.js');
-    const caasUrl = PartnerCards.getCaasUrl(block);
+    const caasUrl = getCaasUrl(block);
     preload(caasUrl);
   });
 }
